# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TodoList API",
    "description": "user post & todoList crud",
    "version": "1.0.0"
  },
  "host": "https://morning-tundra-62406.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "請加上 API Token"
    }
  },
  "paths": {
    "/check": {
      "get": {
        "tags": [
          "Text"
        ],
        "summary": "登入權限測試",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "未授權"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "使用者註冊",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "string"
                    },
                    "name": {
                      "type": "string",
                      "example": "string"
                    },
                    "password": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "註冊成功"
          },
          "422": {
            "description": "註冊失敗"
          }
        }
      }
    },
    "/users/sign_in": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "使用者登入",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "string"
                    },
                    "password": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登入成功"
          },
          "401": {
            "description": "登入失敗"
          }
        }
      }
    },
    "/users/sign_out": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "使用者登出",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "登出成功"
          },
          "401": {
            "description": "登出失敗"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/todos/": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 列表",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "自己的 TODO List"
          },
          "401": {
            "description": "未授權"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "summary": "新增 Todo",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "todo": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "該筆 TODO 資料"
          },
          "401": {
            "description": "未授權"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/todos/{id}": {
      "delete": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 列表",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "自己的 TODO List"
          },
          "401": {
            "description": "未授權"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Todos"
        ],
        "summary": "修改 Todo",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "todo": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改過的 TODO 資料"
          },
          "401": {
            "description": "未授權"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/todos/{id}/toggle": {
      "patch": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 完成/已完成 切換",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo"
          },
          "401": {
            "description": "未授權"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  }
}